requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.3.5
opencv-python-headless==4.12.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.5
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing-extensions==4.15.0
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.6.2
websockets==15.0.1
